Separation of methods. Fixed the stating of the app
