+++ server/Server.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const app = express()
+
+class Server {
+  static init({port = 3000}) {
+    return app.listen(port)
+  }
+}
+
+module.exports = Server;
